Put entity-related-fields in their own component
